+++ public/play/index.php
@@ -31,6 +31,7 @@ declare(strict_types=1);
  */
 
 use Ampache\Module\Application\ApplicationRunner;
+use Ampache\Module\Application\Playback\Play2Action;
 use Ampache\Module\Application\Playback\PlayAction;
 use Ampache\Module\Application\Playback\Play2Action;
 use Nyholm\Psr7Server\ServerRequestCreatorInterface;