+++ docs/CHANGELOG.md
@@ -40,7 +40,7 @@ You can find example Subsonic responses from an official server and Ampache serv
 * webplayer
   * Add a button next to the playlist to allow looping after the last song
   * If you enable playlist loop do not remove previous tracks
-* Database 600025
+* Database 600026
   * Add preference `webplayer_removeplayed`, Remove tracks before the current playlist item in the webplayer when played
   * Drop channel table
   * Add `total_skip` to podcast table
@@ -66,14 +66,14 @@ You can find example Subsonic responses from an official server and Ampache serv
   * Add preference `show_subtitle`, Show Album subtitle on links
   * Add preference `show_original_year`, Show Album original year on links (Separate from use_original_year)
   * Add ui option `show_header_login`, Show the login / registration links in the site header (Separate from simple_user_mode)
-* Config version 66
+  * Add user preference `use_play2`, Use an alternative playback action for streaming if you have issues with playing music
+* Config version 67
   * Drop Channels from config
   * Reset the art_order defaults (replace lastfm with spotify)
   * Set a default `album_art_min_width` and `album_art_min_height` (30px)
   * Add `album_disk` to allow_zip_types
   * Add `fallback_url` for CLI actions which can't detect the URL from web requests
   * Update `additional_genre_delimiters` to `"[/]{2}|[/\\|,;]"` (Split on "//", "_", "/", "\", "|", "," and ";")
-  * Removed `send_full_stream`
   * Update your `encode_args_opus` settings
 * Search
   * Add `album_disk` as a search type (uses album rules)